+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Target Scheduler
 
+## 5.8.1.1 (beta) - 2025-08-07
+* Fix for threading problem in scheduler progress vm add
+
 ## 5.8.1.0 (beta) - 2025-08-07
 * Fixed issue with layout during exposure template editing.
 * Added twilight offset and dither settings to the exposure template profile table.